requirements.txt
playwright
playwright install
pytest-playwright
jproperties
pytest-reporter-html1